♻️ refactor(potential): move the incomplete beta function into its own module
`zhao.py` had grown a ~190-line special-function implementation in the middle
of it, which is neither Zhao-specific nor a potential model. Move it to
`_src/special.py`, alongside `utils.py`/`shape.py`/`jax.py`, leaving `zhao.py`
as just the model.

Fixes a bug found while splitting out the tests: the `b`-derivative was wrong
wherever `b + m` lands exactly on zero, because the pole branch substituted
`log(1/(2w))`, which has no `b` dependence, so autodiff saw a zero tangent.
Integer power-law slopes do hit this (gamma = 2 gives p0 = 0 in Zhao's Eq. 7),
so it would have silently corrupted `jax.grad` with respect to the slopes.

Both branches of that term are now the expansion of one expression,
(e^(s L1) - e^(s L2))/s: the direct form away from the pole, and its series
within `_POLE_BAND` of it, carrying enough terms to be right in both value and
derivative. The L-powers are loop-invariant, so this costs a few FMAs rather
than transcendentals -- an `expm1` formulation was correct too but 30-70%
slower, and the band form is if anything slightly faster than the buggy
original (potential 33 -> 29 ms, hessian 11.5 -> 10.7 ms at 1e5 points).

Also:
- `_mass_fraction_within_r_s` takes the total mass from `betaln` rather than
  evaluating the series at the z = 1 endpoint; it is only reachable for
  q0 > 0, where that is just the complete beta function.
- Tests follow the module split: `tests/unit/potential/test_special.py` for
  the special function (now also covering the endpoints, the a/b derivatives
  against finite differences, second derivatives, and batched-vs-scalar), and
  the analytic-derivative cross-checks move into `test_zhao.py` next to the
  rest of the model's tests.
