+++ src/galax/potential/_src/builtin/zhao.py
@@ -6,10 +6,10 @@ this module refer to that paper.
 
 Everything this model needs -- potential (Eqs. 6-7), enclosed mass (Eq. 15) and
 density (Eq. 1) -- is closed-form in terms of the incomplete beta function
-(Eq. 43), so *no* quantity here is obtained by automatic differentiation:
-`gradient`, `hessian` and `laplacian` are all written analytically below. That
-matters for speed, because the alternative is differentiating through the
-incomplete beta function's series expansion on every force evaluation.
+(Eq. 43, `galax.potential._src.special.incomplete_beta`), so *no* quantity here
+is obtained by automatic differentiation: `gradient`, `hessian` and `laplacian`
+are all written analytically below. That matters for speed, because the
+alternative is differentiating through that function on every force evaluation.
 """
 
 __all__ = ["ZhaoPotential"]
@@ -20,7 +20,7 @@ from typing import final
 
 import equinox as eqx
 import jax
-from jax.custom_derivatives import SymbolicZero
+import jax.scipy.special as jsp
 
 import quaxed.numpy as jnp
 import unxt as u
@@ -33,6 +33,7 @@ from galax.potential._src.base_single import AbstractSinglePotential
 from galax.potential._src.jax import vectorize_method
 from galax.potential._src.params.base import AbstractParameter
 from galax.potential._src.params.field import ParameterField
+from galax.potential._src.special import incomplete_beta
 from galax.potential._src.utils import r_spherical
 
 DimL = u.dimension("length")
@@ -181,202 +182,6 @@ class ZhaoPotential(AbstractSinglePotential):
 
 
 # ===================================================================
-# The incomplete beta function (Zhao Eq. 43)
-#
-# Zhao writes every dynamical quantity in terms of
-#
-#     B(a, b, x) = int_0^x t^(a-1) (1-t)^(b-1) dt              (Eq. 43)
-#
-# the *unregularized* incomplete beta function. `jax.scipy.special` provides
-# only the regularized `betainc`, and reconstructing Eq. 43 as
-# `beta(a, b) * betainc(a, b, x)` is nan whenever `b <= 0` -- exactly the
-# beta <= 3 (infinite total mass) regime -- because the complete beta function
-# diverges there even though the product does not.
-#
-# `jax.scipy.special.hyp2f1` can express it (DLMF 8.17.7) for any `b`, but it
-# is a `lax.while_loop` whose trip count is data-dependent: under `vmap` every
-# lane pays the worst lane's iteration count, and its derivative runs a second
-# such loop. The two fixed-length, geometrically convergent series below cost a
-# predictable ~64 fused multiply-adds instead, and carry an exact O(1)
-# derivative rule.
-#
-# (`nfw/hyp2f1.py` solves the same problem for `gNFWPotential`, but only for
-# the two parameter patterns that model needs -- `a == 1`, or `b == 0`. Zhao's
-# Eqs. 7 and 15 need general `(a, b)`.)
-
-_BZ_NTERMS = 64
-"""Series length. Both series converge like 2^-k, so this is ~1e-19."""
-
-
-def _bz_small_z(a: gt.Sz0, b: gt.Sz0, z: gt.BBtSz0) -> gt.BBtFloatSz0:
-    r"""$B(a, b, z)$ for $z \leq 1/2$, by the defining Taylor series.
-
-    Expanding $(1-t)^{b-1}$ binomially in Eq. 43 and integrating term by term,
-
-    .. math::
-
-        B(a, b, z) = z^a \sum_{k=0}^\infty \frac{(1-b)_k}{k!\,(a+k)} z^k
-
-    The terms fall off like $z^k \leq 2^{-k}$, hence the fixed term count.
-
-    Summed with `jax.lax.scan`, accumulating into the carry: the terms are
-    batched over `z`, so materializing them all at once would cost an
-    ``(*batch, 64)`` temporary and make this memory- rather than flop-bound.
-    """
-
-    def step(
-        carry: tuple[gt.BBtFloatSz0, gt.BBtFloatSz0, gt.Sz0], k: gt.Sz0
-    ) -> tuple[tuple[gt.BBtFloatSz0, gt.BBtFloatSz0, gt.Sz0], None]:
-        total, z_pow, coeff = carry  # coeff = (1-b)_k / k!
-        total = total + coeff / (a + k) * z_pow
-        return (total, z_pow * z, coeff * (k + 1.0 - b) / (k + 1.0)), None
-
-    init = (jnp.zeros_like(z), jnp.ones_like(z), jnp.ones_like(a))
-    (total, _, _), _ = jax.lax.scan(step, init, jnp.arange(_BZ_NTERMS))
-    return z**a * total  # type: ignore[no-any-return]
-
-
-def _bz_large_z(a: gt.Sz0, b: gt.Sz0, z: gt.BBtSz0) -> gt.BBtFloatSz0:
-    r"""$B(a, b, z)$ for $z > 1/2$, by reflecting about $t = 1/2$.
-
-    Substituting $t = 1-u$ in Eq. 43 and splitting the range at $u = 1/2$,
-
-    .. math::
-
-        B(a, b, z) = \int_{w}^{1} u^{b-1}(1-u)^{a-1} du
-                   = B(a, b, 1/2) + \int_w^{1/2} u^{b-1}(1-u)^{a-1} du
-
-    with $w = 1 - z$. Expanding $(1-u)^{a-1}$ binomially (legitimate since
-    $u \leq 1/2$ on the remaining range) and integrating term by term,
-
-    .. math::
-
-        B(a, b, z) = B(a, b, 1/2)
-            + \sum_{m=0}^\infty \frac{(1-a)_m}{m!}
-              \frac{(1/2)^{b+m} - w^{b+m}}{b+m}
-
-    where the $b + m \to 0$ term is $\ln(1/(2w))$. Both pieces converge like
-    $2^{-m}$.
-
-    This branch is what makes $b \leq 0$ work: it never forms the complete beta
-    function $B(a, b)$, which is what diverges there, while keeping the genuine
-    $z \to 1$ divergence of $B(a, b, z)$ itself exact (as $w^b$, or $-\ln w$
-    when $b = 0$).
-
-    Summed with `jax.lax.scan` for the same reason as `_bz_small_z`.
-    """
-    w = 1.0 - z
-    log_half_over_w = jnp.log(0.5 / w)
-
-    def step(
-        carry: tuple[gt.BBtFloatSz0, gt.BBtFloatSz0, gt.Sz0, gt.Sz0], m: gt.Sz0
-    ) -> tuple[tuple[gt.BBtFloatSz0, gt.BBtFloatSz0, gt.Sz0, gt.Sz0], None]:
-        total, w_pow, half_pow, coeff = carry  # coeff = (1-a)_m / m!
-        # ((1/2)^(b+m) - w^(b+m)) / (b+m), whose b+m -> 0 limit is log(1/(2w)).
-        bm = b + m
-        is_pole = jnp.abs(bm) < 1e-12
-        bm_safe = jnp.where(is_pole, 1.0, bm)
-        term = jnp.where(is_pole, log_half_over_w, (half_pow - w_pow) / bm_safe)
-        total = total + coeff * term
-        carry = (total, w_pow * w, half_pow * 0.5, coeff * (m + 1.0 - a) / (m + 1.0))
-        return carry, None
-
-    init = (jnp.zeros_like(w), w**b, 0.5**b, jnp.ones_like(a))
-    (total, _, _, _), _ = jax.lax.scan(step, init, jnp.arange(_BZ_NTERMS))
-    return _bz_small_z(a, b, jnp.full_like(w, 0.5)) + total  # type: ignore[no-any-return]
-
-
-def _incomplete_beta_impl(a: gt.Sz0, b: gt.Sz0, z: gt.BBtSz0) -> gt.BBtFloatSz0:
-    r"""$B(a, b, z)$ (Zhao Eq. 43) for $a > 0$, any real $b$, $z \in [0, 1)$.
-
-    Examples
-    --------
-    >>> import jax.numpy as jnp
-    >>> import jax.scipy.special as jsp
-    >>> from galax.potential._src.builtin.zhao import incomplete_beta
-
-    It agrees with the regularized incomplete beta function wherever that is
-    defined, including close to the $z \to 1$ endpoint:
-
-    >>> a, b = 2.0, 1.5
-    >>> z = jnp.asarray([0.3, 0.999])
-    >>> bool(jnp.allclose(incomplete_beta(a, b, z),
-    ...                   jsp.beta(a, b) * jsp.betainc(a, b, z)))
-    True
-
-    But unlike that product it stays finite for ``b <= 0``, where the complete
-    beta function diverges:
-
-    >>> incomplete_beta(2.0, 0.0, jnp.asarray(0.5))
-    Array(0.19314718, dtype=float64)
-
-    >>> jsp.beta(2.0, 0.0) * jsp.betainc(2.0, 0.0, jnp.asarray(0.5))
-    Array(nan, dtype=float64)
-
-    """
-    z = jnp.asarray(z)
-    # Both branches are evaluated, so clamp each one's input into the range
-    # where it is well behaved; `where` then discards the unused value.
-    return jnp.where(  # type: ignore[no-any-return]
-        z <= 0.5,
-        _bz_small_z(a, b, jnp.minimum(z, 0.5)),
-        _bz_large_z(a, b, jnp.maximum(z, 0.5)),
-    )
-
-
-@jax.custom_jvp
-def incomplete_beta(a: gt.Sz0, b: gt.Sz0, z: gt.BBtSz0) -> gt.BBtFloatSz0:
-    """See `_incomplete_beta_impl` for the definition and examples.
-
-    The `jax.custom_jvp` gives the `z`-derivative in O(1) -- it is just the
-    Eq. 43 integrand evaluated at the endpoint (Leibniz) -- instead of
-    differentiating through a 64-term series. It is a `custom_jvp`, not a
-    `custom_vjp`, so that `jax.hessian`'s `jacfwd(jacrev(...))` still works for
-    anything downstream that has not been given an analytic form.
-    """
-    return _incomplete_beta_impl(a, b, z)
-
-
-@ft.partial(incomplete_beta.defjvp, symbolic_zeros=True)
-def _incomplete_beta_jvp(
-    primals: tuple[gt.Sz0, gt.Sz0, gt.BBtSz0],
-    tangents: tuple[
-        gt.Sz0 | SymbolicZero, gt.Sz0 | SymbolicZero, gt.BBtSz0 | SymbolicZero
-    ],
-) -> tuple[gt.BBtFloatSz0, gt.BBtFloatSz0]:
-    a, b, z = primals
-    a_dot, b_dot, z_dot = tangents
-
-    primal_out = _incomplete_beta_impl(a, b, z)
-    tangent_out = jnp.zeros_like(primal_out)
-
-    # d/dz B(a, b, z) = z^(a-1) (1-z)^(b-1): the Eq. 43 integrand at z.
-    if not isinstance(z_dot, SymbolicZero):
-        tangent_out = tangent_out + z ** (a - 1.0) * (1.0 - z) ** (b - 1.0) * z_dot
-
-    # The a/b tangents are only needed when the power-law indices are
-    # themselves differentiated; there is no cheap closed form, so fall back to
-    # autodiff of the series. Skipped entirely in the common case, where the
-    # differentiation is with respect to position at fixed alpha/beta/gamma.
-    a_zero, b_zero = isinstance(a_dot, SymbolicZero), isinstance(b_dot, SymbolicZero)
-    if not (a_zero and b_zero):
-        _, ab_tangent = jax.jvp(
-            lambda aa, bb: _incomplete_beta_impl(aa, bb, z),
-            (a, b),
-            (
-                jnp.zeros_like(a) if a_zero else a_dot,
-                jnp.zeros_like(b) if b_zero else b_dot,
-            ),
-        )
-        tangent_out = tangent_out + ab_tangent
-
-    return primal_out, tangent_out
-
-
-incomplete_beta = jax.jit(incomplete_beta)  # type: ignore[assignment]
-
-
-# ===================================================================
 # Model internals
 
 
@@ -414,15 +219,16 @@ def _r_to_chi(p: gt.Params, r: gt.BBtSz0, /) -> gt.BBtFloatSz0:
 
 @ft.partial(jax.jit)
 def _mass_fraction_within_r_s(p: gt.Params, /) -> gt.FloatSz0:
-    r"""$M(r_s) / M_{tot}$, from Eq. 15 evaluated at $\chi = 1/2$ and $\chi = 1$.
+    r"""$M(r_s) / M_{tot}$, i.e. Eq. 15 at $\chi = 1/2$ over Eq. 15 at $\chi = 1$.
 
-    Only used by `ZhaoPotential.from_m_tot`, which requires $\beta > 3$ so that
-    $M_{tot} = M(\infty)$ is finite.
+    Only used by `ZhaoPotential.from_m_tot`, which requires $\beta > 3$, i.e.
+    $q_0 > 0$: both that $M_{tot} = M(\infty)$ is finite, and that the total
+    $B(a, q_0, 1)$ below is the ordinary complete beta function.
     """
     c0, _, q0 = _cpq(p["alpha"], p["beta"], p["gamma"])
     a = c0 - q0
-    total = incomplete_beta(a, q0, jnp.asarray(1.0))  # B(a, q0), q0 > 0 here
-    return incomplete_beta(a, q0, jnp.asarray(0.5)) / total
+    total = jnp.exp(jsp.betaln(a, q0))  # B(a, q0, 1) = B(a, q0), for q0 > 0
+    return incomplete_beta(a, q0, jnp.asarray(0.5)) / total  # type: ignore[no-any-return]
 
 
 @ft.partial(jax.jit)
